+++ src/components/StackedAreaChart/hooks/tests/use-y-scale.test.tsx
@@ -12,6 +12,7 @@ jest.mock('d3-scale', () => ({
     scale.range = (range: any) => (range ? scale : range);
     scale.domain = (domain: any) => (domain ? scale : domain);
     scale.nice = () => scale;
+    scale.copy = () => scale;
     return scale;
   }),
 }));
@@ -59,6 +60,7 @@ describe('useYScale', () => {
       scale.range = (range: any) => (range ? scale : range);
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -87,6 +89,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -115,6 +118,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -152,6 +156,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -180,6 +185,7 @@ describe('useYScale', () => {
       scale.range = rangeSpy;
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -206,6 +212,7 @@ describe('useYScale', () => {
       scale.range = (range: any) => (range ? scale : range);
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });